+++ app/src/main/java/ca/pkay/rcloneexplorer/Fragments/RemotesFragment.java
@@ -6,7 +6,6 @@ import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
 import android.support.v4.app.Fragment;
 import android.support.v4.app.FragmentActivity;
-import android.support.v7.app.AppCompatActivity;
 import android.support.v7.widget.LinearLayoutManager;
 import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
@@ -43,6 +42,9 @@ public class RemotesFragment extends Fragment {
     @Override
     public void onCreate(@Nullable Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
+        if (getContext() == null) {
+            return;
+        }
 
         ((FragmentActivity) context).setTitle(getString(R.string.remotes_toolbar_title));
         rclone = new Rclone(getContext());