+++ app/src/main/assets/changelog.md
@@ -1,5 +1,15 @@
+### 1.2.5
+* **New:** Run long running tasks in a Service
+* **New:** File search
+* **Fix:** Caching of directory content
+* **Fix:** Bugs
+
+***
+
 ### 1.2.4
 * **Update:** App shortcut icons are now adaptive and with color
+
+***
 
 ### 1.2.3
 * **New:** Files can be shared with Rclone Explorer